feat: payments lazy loading
