+++ apps/Conduit-UI/src/pages/payments/transactions.tsx
@@ -1,252 +1,20 @@
-import React, { ReactElement, useCallback, useEffect, useState } from 'react';
-import { useAppDispatch, useAppSelector } from '../../redux/store';
-import { EmailUI } from '../../models/emails/EmailModels';
-import { TextField, InputAdornment, Box } from '@mui/material';
-import SearchIcon from '@mui/icons-material/Search';
-import useDebounce from '../../hooks/useDebounce';
+import React, { ReactElement } from 'react';
 import PaymentsLayout from '../../components/navigation/InnerLayouts/paymentsLayout';
-import { asyncDeleteTransactions, asyncGetTransactions } from '../../redux/slices/paymentsSlice';
-import { Transaction } from '../../models/payments/PaymentsModels';
-import ViewTransaction from '../../components/payments/ViewTransaction';
-import {
-  SideDrawerWrapper,
-  DataTable,
-  ConfirmationDialog,
-  TableContainer,
-  TableActionsContainer,
-} from '@conduitplatform/ui-components';
+import dynamic from 'next/dynamic';
+import ScaleLoader from 'react-spinners/ScaleLoader';
+
+const PaymentsTransactions = dynamic(
+  () => import('../../components/payments/PaymentsTransactions'),
+  {
+    loading: () => (
+      <ScaleLoader speedMultiplier={3} color={'#07D9C4'} loading={true} height={21} width={4} />
+    ),
+  }
+);
 
 const Transactions = () => {
-  const dispatch = useAppDispatch();
-
-  const originalTransactionState = {
-    _id: '',
-    userId: '',
-    provider: '',
-    product: '',
-    quantity: 0,
-    data: {},
-    updatedAt: '',
-    createdAt: '',
-  };
-  const [skip, setSkip] = useState<number>(0);
-  const [limit, setLimit] = useState<number>(25);
-  const [page, setPage] = useState<number>(0);
-  const [search, setSearch] = useState<string>('');
-  const [sort, setSort] = useState<{ asc: boolean; index: string | null }>({
-    asc: false,
-    index: null,
-  });
-  const [openDeleteTransactions, setOpenDeleteTransactions] = useState<boolean>(false);
-  const [selectedTransactions, setSelectedTransactions] = useState<string[]>([]);
-  const [drawer, setDrawer] = useState<boolean>(false);
-  const [selectedTransaction, setSelectedTransaction] =
-    useState<Transaction>(originalTransactionState);
-  const debouncedSearch: string = useDebounce(search, 500);
-
-  useEffect(() => {
-    dispatch(
-      asyncGetTransactions({
-        skip,
-        limit,
-        search: debouncedSearch,
-      })
-    );
-  }, [dispatch, limit, skip, debouncedSearch]);
-
-  const { transactions, count } = useAppSelector(
-    (state) => state.paymentsSlice.data.transactionData
-  );
-
-  const handleClose = () => {
-    setDrawer(false);
-    setSelectedTransaction(originalTransactionState);
-    setOpenDeleteTransactions(false);
-  };
-
-  const handlePageChange = (event: React.MouseEvent<HTMLButtonElement> | null, val: number) => {
-    if (val > page) {
-      setPage(page + 1);
-      setSkip(skip + limit);
-    } else {
-      setPage(page - 1);
-      setSkip(skip - limit);
-    }
-  };
-
-  const handleLimitChange = (value: number) => {
-    setLimit(value);
-    setSkip(0);
-    setPage(0);
-  };
-
-  const getCustomersCallBack = useCallback(() => {
-    dispatch(asyncGetTransactions({ skip, limit, search }));
-  }, [dispatch, limit, skip, search]);
-
-  useEffect(() => {
-    setSkip(0);
-    setPage(0);
-    setLimit(25);
-  }, [debouncedSearch]);
-
-  const handleDeleteTitle = (transaction: Transaction) => {
-    if (selectedTransaction._id === '') {
-      return 'Delete selected transactions';
-    }
-    return `Delete transaction ${transaction._id}`;
-  };
-
-  const handleDeleteDescription = (transaction: Transaction) => {
-    if (selectedTransaction._id === '') {
-      return 'Are you sure you want to delete the selected treansactions?';
-    }
-    return `Are you sure you want to delete ${transaction._id}? `;
-  };
-  const deleteButtonAction = () => {
-    if (openDeleteTransactions && selectedTransaction._id == '') {
-      const params = {
-        ids: selectedTransactions,
-        getTransactions: getCustomersCallBack,
-      };
-      dispatch(asyncDeleteTransactions(params));
-    } else {
-      const params = {
-        ids: [`${selectedTransaction._id}`],
-        getTransactions: getCustomersCallBack,
-      };
-      dispatch(asyncDeleteTransactions(params));
-    }
-    setOpenDeleteTransactions(false);
-    setSelectedTransaction(originalTransactionState);
-    setSelectedTransactions([]);
-  };
-
-  const handleAction = (action: { title: string; type: string }, data: any) => {
-    const currentTransaction = transactions?.find((transaction) => transaction._id === data._id);
-
-    if (currentTransaction !== undefined) {
-      if (action.type === 'view') {
-        setSelectedTransaction(currentTransaction);
-        setDrawer(true);
-      }
-    }
-  };
-
-  const handleSelect = (id: string) => {
-    const newSelectedTransactions = [...selectedTransactions];
-
-    if (selectedTransactions.includes(id)) {
-      const index = newSelectedTransactions.findIndex((newId) => newId === id);
-      newSelectedTransactions.splice(index, 1);
-    } else {
-      newSelectedTransactions.push(id);
-    }
-    setSelectedTransactions(newSelectedTransactions);
-  };
-
-  const handleSelectAll = (data: EmailUI[]) => {
-    if (selectedTransactions.length === transactions.length) {
-      setSelectedTransactions([]);
-      return;
-    }
-    const newSelectedTransactions = data.map((item: EmailUI) => item._id);
-    setSelectedTransactions(newSelectedTransactions);
-  };
-
-  const toDelete = {
-    title: 'Delete',
-    type: 'delete',
-  };
-
-  const toView = {
-    title: 'View',
-    type: 'view',
-  };
-
-  const actions = [toView, toDelete];
-
-  const formatData = (data: Transaction[]) => {
-    return data.map((u) => {
-      return {
-        _id: u._id,
-        provider: u.provider,
-        product: u.product,
-        quantity: u.quantity,
-        'Updated At': u.updatedAt,
-      };
-    });
-  };
-
-  const headers = [
-    { title: '_id', sort: '_id' },
-    { title: 'Provider', sort: 'provider' },
-    { title: 'Product', sort: 'product' },
-    { title: 'Quantity', sort: 'quantity' },
-    { title: 'Updated At', sort: 'updatedAt' },
-  ];
-
-  return (
-    <Box>
-      <TableActionsContainer>
-        {count >= 0 && (
-          <TextField
-            size="small"
-            variant="outlined"
-            name="Search"
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            label="Find transaction"
-            InputProps={{
-              startAdornment: (
-                <InputAdornment position="start">
-                  <SearchIcon />
-                </InputAdornment>
-              ),
-            }}
-          />
-        )}
-      </TableActionsContainer>
-      <TableContainer
-        handlePageChange={handlePageChange}
-        limit={limit}
-        handleLimitChange={handleLimitChange}
-        page={page}
-        count={count}
-        noData={!transactions.length ? 'transactions' : undefined}>
-        <DataTable
-          sort={sort}
-          setSort={setSort}
-          headers={headers}
-          dsData={formatData(transactions)}
-          handleSelect={handleSelect}
-          handleSelectAll={handleSelectAll}
-          actions={actions}
-          handleAction={handleAction}
-          selectedItems={selectedTransactions}
-        />
-      </TableContainer>
-      <SideDrawerWrapper
-        title={'Transaction overview'}
-        open={drawer}
-        closeDrawer={() => handleClose()}
-        width={750}>
-        <Box>
-          <ViewTransaction transaction={selectedTransaction} />
-        </Box>
-      </SideDrawerWrapper>
-      <ConfirmationDialog
-        open={openDeleteTransactions}
-        handleClose={handleClose}
-        title={handleDeleteTitle(selectedTransaction)}
-        description={handleDeleteDescription(selectedTransaction)}
-        buttonAction={deleteButtonAction}
-        buttonText={'Delete'}
-      />
-    </Box>
-  );
+  return <PaymentsTransactions />;
 };
-
 Transactions.getLayout = function getLayout(page: ReactElement) {
   return <PaymentsLayout>{page}</PaymentsLayout>;
 };
